pyproject.toml
[project]
name = "anthias"
version = "0.20.4"
description = "The world's most popular open source digital signage project."
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
server = [
    "certifi==2026.2.25",
    "netifaces==0.11.0",
    "cffi==2.0.0",
    "channels==4.2.0",
    "configparser==7.2.0",
    "cryptography==46.0.5",
    "Cython==3.2.4",
    "daphne==4.1.2",
    "Django==4.2.29",
    "djangorestframework==3.16.1",
    "django-dbbackup==4.2.1",
    "drf-spectacular==0.29.0",
    "future==1.0.0",
    "hurry.filesize==0.9",
    "importlib-metadata==8.7.1",
    "Jinja2==3.1.6",
    "jsonschema==4.26.0",
    "Mako==1.3.10",
    "psutil==7.2.2",
    "pyasn1==0.6.2",
    "pydbus==0.6.0",
    "pyOpenSSL==25.3.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2026.1.post1",
    "PyYAML==6.0.3",
    "redis==6.2.0",
    "requests[security]==2.32.5",
    "tenacity==9.1.4",
    "sh==2.2.2",
    "six==1.17.0",
    "urllib3==2.6.3",
    "websockets==15.0.1",
    "wheel==0.46.3",
    "whitenoise==6.9.0",
    "yt-dlp==2026.3.3",
]
viewer = [
    "certifi==2026.2.25",
    "configparser==7.2.0",
    "cryptography==46.0.5",
    "Cython==3.2.4",
    "Django==4.2.29",
    "django-dbbackup==4.2.1",
    "drf-spectacular==0.29.0",
    "future==1.0.0",
    "idna==3.11",
    "Jinja2==3.1.6",
    "jsonschema==4.26.0",
    "netifaces==0.11.0",
    "pydbus==0.6.0",
    "python-dateutil==2.9.0.post0",
    "python-vlc==3.0.21203",
    "pytz==2026.1.post1",
    "requests[security]==2.32.5",
    "tenacity==9.1.4",
    "sh==2.2.2",
    "uptime==3.0.1",
    "urllib3==2.6.3",
    "websockets==15.0.1",
]
wifi-connect = [
    "Cython==3.2.4",
    "future==1.0.0",
    "netifaces==0.11.0",
]
dev = [
    "future==1.0.0",
    "pep8==1.7.1",
    "pytest==8.3.5",
    "pytest-django==4.10.0",
    "selenium==4.41.0",
    "splinter==0.21.0",
    "time-machine==3.2.0",
]
test = [
    { include-group = "server" },
    { include-group = "dev" },
    { include-group = "viewer" },
]
host = [
    "ansible-core==2.18.3",
    "getmac==0.9.5",
    "netifaces2==0.0.22",
    "requests[security]==2.32.5",
    "tenacity==9.1.4",
]
local = [
    "click==8.3.1",
    "requests==2.32.5",
    "tenacity==9.1.4",
]
dev-host = ["ruff==0.15.5"]
docker-image-builder = [
    "click==8.3.1",
    "jinja2==3.1.6",
    "pygit2==1.19.1",
    "requests==2.32.5",
    "python-on-whales==0.81.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "anthias_django.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
